+++ cumulusci/tasks/bulkdata/mapping_parser.py
@@ -269,7 +269,7 @@ class MappingStep(CCIDictModel):
                 "If you set a `batch_size` you should also set an `api` to `rest` or `bulk`. "
                 "`batch_size` means different things for `rest` and `bulk`. "
                 "Please see the documentation for further details. "
-                "https://cumulusci.readthedocs.io/en/latest/data.html#api-selection"
+                "https://claritisoftware.github.io/CumulusCI/data.html#api-selection"
             )
         else:  # pragma: no cover
             # should not happen